Rename servoMix_options.h to options_servo_mix.h for MPLAB-X projects.
