+++ MatrixPilot/MatrixPilot-udb4.X/nbproject/configurations.xml
@@ -26,11 +26,11 @@
         <itemPath>../../Config/options_auav3.h</itemPath>
         <itemPath>../../Config/osd_config.h</itemPath>
         <itemPath>../../Config/ports_config.h</itemPath>
-        <itemPath>../../Config/servoMix_options.h</itemPath>
         <itemPath>../../Config/waypoints.h</itemPath>
         <itemPath>../../Config/flightplan-logo.h</itemPath>
         <itemPath>../../Config/options_magnetometer.h</itemPath>
         <itemPath>../../Config/options_airspeed.h</itemPath>
+        <itemPath>../../Config/options_servo_mix.h</itemPath>
       </logicalFolder>
       <logicalFolder name="libDCM" displayName="libDCM" projectFiles="true">
         <itemPath>../../libDCM/dcmTypes.h</itemPath>
