+++ Tools/flight_analyzer/flan_config
@@ -1,0 +1,56 @@
+(i__main__
+flan_options
+p0
+(dp1
+S'altitude_correction'
+p2
+I10
+sS'gps_delay_correction'
+p3
+I0
+sS'telemetry_filename'
+p4
+S'/Users/phollands/Desktop/uav/flight_analysis/flight_logs/alpha_139_UDB5_White_Tip/LOG000016/LOG00375.TXT'
+p5
+sS'telemetry_type'
+p6
+S'SERIAL_UDB_EXTRA'
+p7
+sS'GE_selector'
+p8
+I1
+sS'loglevel'
+p9
+I0
+sS'graph'
+p10
+I0
+sS'GE_filename_kml'
+p11
+S'/Users/phollands/Desktop/uav/flight_analysis/flight_logs/alpha_139_UDB5_White_Tip/LOG000016/LOG00375.kml'
+p12
+sS'CSV_filename'
+p13
+S'/Users/phollands/Desktop/uav/flight_analysis/flight_logs/alpha_139_UDB5_White_Tip/LOG000016/LOG00375.csv'
+p14
+sS'waypoint_filename'
+p15
+S'/Users/phollands/Desktop/uav/flight_analysis/flight_logs/alpha_139_UDB5_White_Tip/LOG000016/Config/waypoints.h'
+p16
+sS'CSV_selector'
+p17
+I1
+sS'relocate'
+p18
+I0
+sS'GE_filename'
+p19
+S'/Users/phollands/Desktop/uav/flight_analysis/flight_logs/alpha_139_UDB5_White_Tip/LOG000016/LOG00375.kmz'
+p20
+sS'waypoint_selector'
+p21
+I1
+sS'telemetry_selector'
+p22
+I1
+sb.
